refactor: restructure Career section by moving to separate files and improving component organization
